requirements.txt
argparse
spotipy
requests
python-dotenv